Moved form parsing to its own package
