+++ collect_app/src/main/java/org/odk/collect/android/formmanagement/LocalFormUseCases.kt
@@ -3,6 +3,8 @@ package org.odk.collect.android.formmanagement
 import android.database.SQLException
 import org.javarosa.xform.parse.XFormParser
 import org.odk.collect.android.application.Collect
+import org.odk.collect.android.formmanagement.metadata.FormMetadata
+import org.odk.collect.android.formmanagement.metadata.FormMetadataParser
 import org.odk.collect.android.utilities.FileUtils
 import org.odk.collect.androidshared.utils.Validator
 import org.odk.collect.forms.Form
