+++ collect_app/src/main/java/org/odk/collect/android/formmanagement/metadata/FormMetadata.kt
@@ -1,0 +1,12 @@
+package org.odk.collect.android.formmanagement.metadata
+
+data class FormMetadata(
+    val title: String?,
+    val id: String?,
+    val version: String?,
+    val submissionUri: String?,
+    val base64RsaPublicKey: String?,
+    val autoDelete: String?,
+    val autoSend: String?,
+    val geometryXPath: String?
+)
